refactor: 컨트롤러 클래스 세분화 및 api uri 조정
