+++ src/main/java/roomescape/controller/admin/AdminReservationController.java
@@ -1,0 +1,67 @@
+package roomescape.controller.admin;
+
+import java.time.LocalDate;
+import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.AdminReservationRequest;
+import roomescape.dto.ReservationResponse;
+import roomescape.service.AdminReservationService;
+import roomescape.service.ReservationService;
+
+@RestController
+@RequestMapping("/admin/reservations")
+public class AdminReservationController {
+
+    //TODO : 둘을 어떻게 합치거나 분리할지 모르겠다.
+    private final AdminReservationService adminReservationService;
+    private final ReservationService reservationService;
+
+    public AdminReservationController(AdminReservationService adminReservationService,
+                                      ReservationService reservationService) {
+        this.adminReservationService = adminReservationService;
+        this.reservationService = reservationService;
+    }
+
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
+    public List<ReservationResponse> readReservations(
+            @RequestParam(required = false) Long themeId,
+            @RequestParam(required = false) Long memberId,
+            @RequestParam(required = false) LocalDate dateFrom,
+            @RequestParam(required = false) LocalDate dateTo
+    ) {
+        if (themeId != null
+                || memberId != null
+                || dateFrom != null
+                || dateTo != null
+        ) {
+            return reservationService.searchReservations(themeId, memberId, dateFrom, dateTo);
+        }
+        return reservationService.findAllReservations();
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public ReservationResponse createReservation(
+            @RequestBody AdminReservationRequest request
+    ) {
+        return adminReservationService.createReservation(request);
+    }
+
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteReservation(
+            @PathVariable long id
+    ) {
+        reservationService.deleteReservation(id);
+    }
+}
