+++ src/main/java/roomescape/controller/admin/AdminReservationTimeController.java
@@ -1,0 +1,49 @@
+package roomescape.controller.admin;
+
+import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.ReservationTimeRequest;
+import roomescape.dto.ReservationTimeResponse;
+import roomescape.service.ReservationTimeService;
+
+@RestController
+@RequestMapping("/admin/times")
+public class AdminReservationTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public AdminReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
+    public List<ReservationTimeResponse> readTimes(
+    ) {
+        return reservationTimeService.findAllTimes();
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public ReservationTimeResponse createTime(
+            @RequestBody ReservationTimeRequest request
+    ) {
+        return reservationTimeService.createTime(request);
+    }
+
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteTime(
+            @PathVariable Long id
+    ) {
+        reservationTimeService.deleteTime(id);
+    }
+}
