+++ src/main/java/roomescape/controller/admin/AdminThemeController.java
@@ -1,0 +1,48 @@
+package roomescape.controller.admin;
+
+import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.ThemeRequest;
+import roomescape.dto.ThemeResponse;
+import roomescape.service.ThemeService;
+
+@RestController
+@RequestMapping("/admin/themes")
+public class AdminThemeController {
+
+    private final ThemeService themeService;
+
+    public AdminThemeController(ThemeService themeService) {
+        this.themeService = themeService;
+    }
+
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
+    public List<ThemeResponse> readThemes() {
+        return themeService.findAllThemes();
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public ThemeResponse createTheme(
+            @RequestBody ThemeRequest request
+    ) {
+        return themeService.createTheme(request);
+    }
+
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteTheme(
+            @PathVariable Long id
+    ) {
+        themeService.deleteTheme(id);
+    }
+}
