+++ src/main/java/roomescape/controller/common/AuthController.java
@@ -1,43 +1,28 @@
-package roomescape.controller;
+package roomescape.controller.common;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.config.LoginMember;
 import roomescape.dto.AuthorizationResponse;
 import roomescape.dto.LoginRequest;
-import roomescape.dto.MemberResponse;
-import roomescape.dto.SignupRequest;
 import roomescape.entity.AccessToken;
 import roomescape.service.MemberService;
 
 @RestController
-public class MemberController {
+@RequestMapping("/auth")
+public class AuthController {
 
     private final MemberService memberService;
 
-    public MemberController(MemberService memberService) {
+    public AuthController(MemberService memberService) {
         this.memberService = memberService;
-    }
-
-    @GetMapping("/members")
-    @ResponseStatus(HttpStatus.OK)
-    public List<MemberResponse> readMembers() {
-        return memberService.findAllMembers();
-    }
-
-    @PostMapping("/members")
-    @ResponseStatus(HttpStatus.CREATED)
-    public MemberResponse createMember(
-            @RequestBody SignupRequest request
-    ) {
-        return memberService.createUser(request);
     }
 
     @PostMapping("/login")
