+++ src/main/java/roomescape/controller/common/MemberController.java
@@ -1,0 +1,38 @@
+package roomescape.controller.common;
+
+import java.util.List;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.dto.MemberResponse;
+import roomescape.dto.SignupRequest;
+import roomescape.service.MemberService;
+
+@RestController
+@RequestMapping("/members")
+public class MemberController {
+
+    private final MemberService memberService;
+
+    public MemberController(MemberService memberService) {
+        this.memberService = memberService;
+    }
+
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
+    public List<MemberResponse> readMembers() {
+        return memberService.findAllMembers();
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public MemberResponse createMember(
+            @RequestBody SignupRequest request
+    ) {
+        return memberService.createUser(request);
+    }
+}
