+++ src/main/java/roomescape/controller/user/ReservationController.java
@@ -1,15 +1,11 @@
-package roomescape.controller;
+package roomescape.controller.user;
 
-import java.time.LocalDate;
-import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.config.LoginMember;
@@ -25,24 +21,6 @@ public class ReservationController {
 
     public ReservationController(ReservationService reservationService) {
         this.reservationService = reservationService;
-    }
-
-    @GetMapping
-    @ResponseStatus(HttpStatus.OK)
-    public List<ReservationResponse> readReservations(
-            @RequestParam(required = false) Long themeId,
-            @RequestParam(required = false) Long memberId,
-            @RequestParam(required = false) LocalDate dateFrom,
-            @RequestParam(required = false) LocalDate dateTo
-    ) {
-        if (themeId != null
-                || memberId != null
-                || dateFrom != null
-                || dateTo != null
-        ) {
-            return reservationService.searchReservations(themeId, memberId, dateFrom, dateTo);
-        }
-        return reservationService.findAllReservations();
     }
 
     @PostMapping
