+++ src/main/java/roomescape/controller/user/ReservationTimeController.java
@@ -1,20 +1,14 @@
-package roomescape.controller;
+package roomescape.controller.user;
 
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.ReservationTimeAvailableResponse;
-import roomescape.dto.ReservationTimeRequest;
-import roomescape.dto.ReservationTimeResponse;
 import roomescape.service.ReservationTimeService;
 
 @RestController
@@ -27,13 +21,6 @@ public class ReservationTimeController {
         this.reservationTimeService = reservationTimeService;
     }
 
-    @GetMapping
-    @ResponseStatus(HttpStatus.OK)
-    public List<ReservationTimeResponse> readTimes(
-    ) {
-        return reservationTimeService.findAllTimes();
-    }
-
     @GetMapping("/available")
     @ResponseStatus(HttpStatus.OK)
     public List<ReservationTimeAvailableResponse> readAvailableTimes(
@@ -41,21 +28,5 @@ public class ReservationTimeController {
             @RequestParam Long themeId
     ) {
         return reservationTimeService.findAvailableTimes(date, themeId);
-    }
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public ReservationTimeResponse createTime(
-            @RequestBody ReservationTimeRequest request
-    ) {
-        return reservationTimeService.createTime(request);
-    }
-
-    @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteTime(
-            @PathVariable Long id
-    ) {
-        reservationTimeService.deleteTime(id);
     }
 }
