+++ src/main/java/roomescape/controller/user/ThemeController.java
@@ -1,16 +1,11 @@
-package roomescape.controller;
+package roomescape.controller.user;
 
 import java.util.List;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
 import roomescape.service.ThemeService;
 
@@ -34,21 +29,5 @@ public class ThemeController {
     @ResponseStatus(HttpStatus.OK)
     public List<ThemeResponse> readThemeRank() {
         return themeService.findThemeRank();
-    }
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public ThemeResponse createTheme(
-            @RequestBody ThemeRequest request
-    ) {
-        return themeService.createTheme(request);
-    }
-
-    @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteTheme(
-            @PathVariable Long id
-    ) {
-        themeService.deleteTheme(id);
     }
 }
