+++ src/main/java/roomescape/controller/user/UserPageController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.user;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
